Extract Habits.get_current_streak
And extract HabitsTest
