+++ test/habits/habit_test.exs
@@ -1,6 +1,7 @@
 defmodule Habits.HabitTest do
   use Habits.DataCase
 
+  alias Habits.Date, as: DateHelpers
   alias Habits.Habits.Habit
 
   @valid_attrs %{name: "some content", account_id: 42}
@@ -16,40 +17,6 @@ defmodule Habits.HabitTest do
     refute changeset.valid?
   end
 
-  describe ".get_current_streak" do
-    test "gets the current_streak before checking in today" do
-      habit = Factory.insert(:habit)
-      Factory.insert(:check_in, habit: habit, date: days_ago(2))
-      Factory.insert(:check_in, habit: habit, date: days_ago(1))
-
-      assert Habit.get_current_streak(habit) == 2
-    end
-
-    test "gets the current_streak after checking in today" do
-      habit = Factory.insert(:habit)
-      Factory.insert(:check_in, habit: habit, date: days_ago(2))
-      Factory.insert(:check_in, habit: habit, date: days_ago(1))
-      Factory.insert(:check_in, habit: habit, date: days_ago(0))
-
-      assert Habit.get_current_streak(habit) == 3
-    end
-
-    test "breaks the current_streak when your previous check-in was 2 days ago" do
-      habit = Factory.insert(:habit)
-      Factory.insert(:check_in, habit: habit, date: days_ago(2))
-      Factory.insert(:check_in, habit: habit, date: days_ago(0))
-
-      assert Habit.get_current_streak(habit) == 1
-    end
-
-    test "zeroes out the current_streak when you haven't checked in in 2 days" do
-      habit = Factory.insert(:habit)
-      Factory.insert(:check_in, habit: habit, date: days_ago(2))
-
-      assert Habit.get_current_streak(habit) == 0
-    end
-  end
-
   describe ".get_longest_streak" do
     test "returns the longest consecutive streak for a habit" do
       habit = Factory.insert(:habit)
@@ -63,7 +30,7 @@ defmodule Habits.HabitTest do
   end
 
   defp days_ago(days) do
-    Habits.Date.today()
-    |> Habits.Date.shift_days(-days)
+    DateHelpers.today()
+    |> DateHelpers.shift_days(-days)
   end
 end
